requirements.txt
altair==5.5.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.9
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.4.1
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docopt==0.6.2
filelock==3.17.0
firebase-admin==6.6.0
flatbuffers==25.2.10
fonttools==4.55.8
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
googletrans==4.0.2
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.5
Js2Py==0.74
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langdetect==1.0.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
narwhals==1.25.0
networkx==3.4.2
numpy==2.1.1
onnxruntime==1.20.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pipwin==0.5.2
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
py-cpuinfo==9.0.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
pyjsparser==2.7.1
PyJWT==2.10.1
pyparsing==3.2.1
pypiwin32==223
PyPrind==2.11.3
pyreadline3==3.5.4
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytz==2025.1
pywin32==308
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
scipy==1.15.1
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
ultralytics==8.3.70
ultralytics-thop==2.0.14
uritemplate==4.1.1
urllib3==2.3.0
watchdog==6.0.0
wheel==0.45.1
xvfbwrapper==0.2.9
